+++ CHANGELOG.md
@@ -13,10 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - If a tuple `{:allow, user_id}` is returned from the `log_in_requested/1` callback, Magic Auth will store the user `ID` in 
     `%Session{user_id: user_id}`. The user will then be automatically loaded by `fetch_magic_auth_session/2` and assigned 
     to `assigns` under the `:current_user` key.
+  - Support for multi tenancy with query prefixes and foreign keys.
 
 ### Breaking changes
 
-  To support automatic user loading feature, a few small changes must be made to your project before updating:
+  To support the automatic user loading feature, a few small changes must be made to your project before updating (only if you want to use this functionality):
 
   1. Add `user_id` column on `magic_auth_sessions` table.
   ```elixir
@@ -28,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   2. Add your user schema to the Magic Auth configuration on `config/config.exs`:
   ```elixir
   config :magic_auth,
-    user_schema: MyApp.User
+    get_user: %MyApp.Accounts.get_user_by_id/1
   ```
 
   DONE!